fixes BenchmarkingTests.java to allow them to be skipped silently

+++ src/test/java/BenchmarkingTests.java
@@ -13,12 +13,14 @@ import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.concurrent.*;
 
+
 /**
  * Benchmarks end-to-end record and replay timing characteristics.
  */
 // allows before all annotation to be non-static, see https://docs.junit.org/current/api/org.junit.jupiter.api/org/junit/jupiter/api/TestInstance.Lifecycle.html
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class BenchmarkingTests {
+    boolean benchmarking = false;
     // TODO: test fast keyboard inputs (and mouse!)
     private final String predeterminedEvents = """
 START KEY EVENTS
@@ -51,6 +53,9 @@ EOF
 
     @BeforeAll
     void init() throws RuntimeException, IOException {
+        if (!benchmarking) {
+            return;
+        }
         // before all the tests, setup the file it will read from.
         // check if the tmp dir exists, create if not
         tmpDirFile = new File(tmpDirPath);
@@ -72,6 +77,9 @@ EOF
 
     @AfterAll
     void cleanup() throws IOException {
+        if (!benchmarking) {
+            return;
+        }
         // delete our tmp file and dir
         Files.deleteIfExists(Paths.get(tmpRecorderOutPath));
         Files.deleteIfExists(predeterminedEventsFile.toPath());
@@ -80,17 +88,21 @@ EOF
 
     @Test
     void benchmark() throws InterruptedException {
+        if (!benchmarking) {
+            System.out.println("WARNING: Benchmarking is not enabled!");
+            return;
+        }
         File out = new File(tmpRecorderOutPath);
         // so now we need to set up a replayer, feed it the predetermined events, then set up a recorder.
         // latch allows us to countdown to execution, getting pretty perfect execution times
-        Replayer replayer = new Replayer(predeterminedEventsFile.getAbsolutePath(), -1);
+        Replayer replayer = new Replayer(predeterminedEventsFile.getAbsolutePath(), 1);
         Recorder recorder = new Recorder(out, "ESCAPE");
         ExecutorService exec = Executors.newFixedThreadPool(2);
         CountDownLatch latch = new CountDownLatch(1);
 
         exec.submit(() -> {
             try {
-                latch.await();  // 
+                latch.await();  // both wait for signal
             } catch (InterruptedException e) {
                 throw new RuntimeException(e);
             }
@@ -161,7 +173,7 @@ EOF
 
         // some hard limits on timings that we should always pass
         assertTrue(mad < MAD_CUTOFF);
-        assertTrue(mag <= MAG_CUTOFF);
+        assertTrue(mag < MAG_CUTOFF);
 
         // TODO: if were feeling nice, make a custom macro here and feed it to a new replayer that just holds backspace for like a second to get rid of stuff typed above
     }
